ME176C: Revert "pinctrl: baytrail: Do not add all GPIOs to IRQ domain"
Causes problems with Bluetooth. We use one of the GPIOs as interrupt,
to wake up the Bluetooth chip when it's in a low power mode. However,
the GPIO is not configured correctly in the hardware configuration.

With that commit, it can no longer be used as IRQ, so just revert it
for now.

This reverts commit 49c03096263871a68c9dea3e86b7d1e163d2fba8.